Move hash to public, to build logging sample app
