+++ src/consensus/aft/test/driver.cpp
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #include "driver.h"
 
-#include "ds/hash.h"
+#include "ccf/ds/hash.h"
 
 #include <cassert>
 #include <fstream>
