+++ src/ds/champ_map.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "ccf/ccf_assert.h"
-#include "ds/hash.h"
+#include "ccf/ds/hash.h"
 #include "ds/map_serializers.h"
 
 #include <algorithm>
