+++ src/ds/test/hash.cpp
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
-#include "ds/hash.h"
+#include "ccf/ds/hash.h"
 
 #include "ccf/ds/siphash.h"
 #include "siphash_known_hashes.h"
