+++ src/ds/test/hash_bench.cpp
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 
 #include "ccf/byte_vector.h"
-#include "ds/hash.h"
+#include "ccf/ds/hash.h"
 
 #define PICOBENCH_IMPLEMENT_WITH_MAIN
 #include <picobench/picobench.hpp>
